cleanups

+++ tutorials/pathtracer/pathtracer_device.cpp
@@ -1178,7 +1178,7 @@ RTCScene convertScene(ISPCScene* scene_in)
   /* no instancing */
   else
   {
-    for (size_t i=0; i<scene_in->numGeometries; i++)
+    for (unsigned int i=0; i<scene_in->numGeometries; i++)
     {
       ISPCGeometry* geometry = scene_in->geometries[i];
       if (geometry->type == SUBDIV_MESH) {
